ENH: Adds superbuild process.
This will ease building of LEWOS by automating the build of
vrpn, LEWOS_sim_lib and LEWOS_xlate_lib, building all the necessary
libraries.

TODO: Use a separate directory for the main MACRO executable and
tester.  This way the superbuild process does not need to be the
only build process, and changes can be made to the executable without
updating/rebuilding the other libraries.